int to double casting

+++ lib/pages/adminPage/adminDashBoard.dart
@@ -156,7 +156,7 @@ Widget buildCard(
                       Text(
                         "Temprature:  ",
                         style: TextStyle(
-                          color: int.parse(temprature) < 37.2
+                          color: double.parse(temprature) < 37.2
                               ? Colors.black54
                               : Colors.red,
                           fontWeight: FontWeight.bold,
@@ -182,7 +182,7 @@ Widget buildCard(
                       AutoSizeText(
                         temprature + '°C',
                         style: TextStyle(
-                          color: int.parse(temprature) < 37.2
+                          color: double.parse(temprature) < 37.2
                               ? Colors.black54
                               : Colors.red,
                           fontWeight: FontWeight.bold,
