+++ lib/pages/studentPage/userDashboard.dart
@@ -113,7 +113,7 @@ Widget buildCard(
                       Text(
                         "Temprature:  ",
                         style: TextStyle(
-                          color: int.parse(temprature) < 37.2
+                          color: double.parse(temprature) < 37.2
                               ? Colors.black54
                               : Colors.red,
                           fontWeight: FontWeight.bold,
@@ -138,7 +138,7 @@ Widget buildCard(
                       AutoSizeText(
                         temprature + '°C',
                         style: TextStyle(
-                          color: int.parse(temprature) < 37.2
+                          color: double.parse(temprature) < 37.2
                               ? Colors.black54
                               : Colors.red,
                           fontWeight: FontWeight.bold,
